Fix for http://freedesktop.org/bugzilla/show_bug.cgi?id=1489 - Use |long|     instead of |int| for BIGREQUESTS commandline option.

+++ os/utils.c
@@ -1,4 +1,4 @@
-/* $XdotOrg: xc/programs/Xserver/os/utils.c,v 1.7 2004/09/14 00:51:25 gisburn Exp $ */
+/* $XdotOrg: xc/programs/Xserver/os/utils.c,v 1.8 2004/09/24 02:11:14 gisburn Exp $ */
 /* $Xorg: utils.c,v 1.5 2001/02/09 02:05:24 xorgcvs Exp $ */
 /*
 
@@ -1008,11 +1008,11 @@ ProcessCommandLine(int argc, char *argv[])
 	    defaultBackingStore = WhenMapped;
         else if ( strcmp( argv[i], "-maxbigreqsize") == 0) {
              if(++i < argc) {
-                 int reqSizeArg = atoi(argv[i]);
+                 long reqSizeArg = atol(argv[i]);
 
                  /* Request size > 128MB does not make much sense... */
-                 if( reqSizeArg > 0 && reqSizeArg < 128 ) {
-                     maxBigRequestSize = (reqSizeArg * 1048576) - 1;
+                 if( reqSizeArg > 0L && reqSizeArg < 128L ) {
+                     maxBigRequestSize = (reqSizeArg * 1048576L) - 1L;
                  }
                  else
                  {
